digiKam from trunk : BUGFIX in kioslave :
Marc Espie from OpenBSD team have investigate why album thumbs rendering is broken to the first item into openBSD. There is a portability into kioslave about 'st_size' system definition ('long long' instead  'unsigned long').

Any digikam users have reported a problem with thumbs rendering under Linux. I suspect that we have the same problem than OpenBSD. Please give me a report for that.

CCMAIL: digikam-devel@kde.org

svn path=/trunk/extragear/graphics/digikam/; revision=518981

+++ kioslave/digikamalbums.cpp
@@ -206,7 +206,7 @@ void kio_digikamalbums::special(const QByteArray& data)
         os << albumid;
         os << name;
         os << date;
-        os << stbuf.st_size;
+        os << static_cast<size_t>(stbuf.st_size);
         os << dims;
     }
