+++ kioslave/digikamsearch.cpp
@@ -216,7 +216,7 @@ void kio_digikamsearch::special(const QByteArray& data)
             *os << dirid;
             *os << name;
             *os << date;
-            *os << stbuf.st_size;
+            *os << static_cast<size_t>(stbuf.st_size);
             *os << dims;
 
             count++;
