+++ kioslave/digikamtags.cpp
@@ -193,7 +193,7 @@ void kio_digikamtagsProtocol::special(const QByteArray& data)
         *os << dirid;
         *os << name;
         *os << date;
-        *os << stbuf.st_size;
+        *os << static_cast<size_t>(stbuf.st_size);
         *os << dims;
 
         count++;
